#1969: Fixed the chance calculation in CValueCurveDef::GetLinear using a mix of signed and unsigned variables, leading to weird results (particularly when calculating skill action timer). - Forced CResourceBase::ResourceGetName to print GetPrivateUID as a signed long, to fix negative values being converted to unsigned. - Moved handling of Linux console interaction to a new "UnixTerminal" class. The behaviour has been changed slightly so that characters aren't automatically echoed back to the console, giving us a chance to handle control characters (e.g. backspace) properly without display strange symbols on the screen. - Modified coloured console handling on Linux so that colours are enabled by default on certain terminal types (supposedly ones that should support colour -- can undo if this is not the case). - Added experimental support for using ncurses to handle console interactions on Linux. This seems to work well and looks nice, but unfortunately using it seems to disable the ability to scroll up through the console output so for now I'm leaving it disabled. Add a _USECURSES #define and link to the "ncurses" library to enable this.

+++ 0.56.b/graysvr/CResourceDef.cpp
@@ -131,7 +131,7 @@ int CValueCurveDef::GetLinear( int iSkillPercent ) const
 
 	int iLoVal = m_aiValues[iLoIdx];
 	int iHiVal = m_aiValues[iLoIdx + 1];
-	int iChance = iLoVal + IMULDIV( iHiVal - iLoVal, iSkillPercent, iSegSize );
+	int iChance = iLoVal + IMULDIV( iHiVal - iLoVal, static_cast<int>(iSkillPercent), static_cast<int>(iSegSize) );
 
 	if ( iChance <= 0 )
 		return 0; // less than no chance ?
